xhci: move queued deq segment and pointer from endpoint to ring
With streams we have several stream rings per endpoint.
We might need to set the TR dequeue pointer for each stream of a endpoint
while the endpoint is stopped, and need to store a new dequeue pointer
position for each stream.

Move the stored queued deq pointer from endpoint to endpoint/stream ring

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>