xhci: set TR dequeue pointer for each cancelled stream URB
With streams we have several stream rings per endpoint.
When a endpoint is stopped to cancel URBs all endpoint streams
stop as well.

When several URBs for different streams on the same endpoint should be
cancelled we need to set the TR dequeue pointer for each stream that
stopped on the URB that should be cancelled.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>