usb: xhci: replace custom ring segment list with standard list.h lists
Replace the custom list implementation with the standard list.h list.
The transition is made with minimal function adjustments to maintain
the simplicity and readability of this commit. Functions will be
adjusted/optimized for the new list style in future commits.

The reason for this change is to simplify xhci ring segment by using
a common list system, as well as enabling us to use the features
provided by list.h, such as bilateral list access.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>