usb: xhci: remove struct xhci_ring's pointers to first & last segment
Remove the direct pointers to the first and last segments within
struct 'xhci_ring'. Instead, use 'RING_FIRST_SEG()' and 'RING_LAST_SEG()'
macros. This modification mitigates the risk of having incorrect
first/last segment pointers and the need to update them.

The rationale behind the introduction of xhci specific macros, as opposed
to directly using the list.h macro, is to simplify and declutter the code.
The list.h macro requires the data type and the name of the 'list_head'
variable, which can be cumbersome. By encapsulating these details within
the xhci macro, the code remains simple and clutter free.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>